test: move from `src` to `test` folder
The project structure become more organized
Use `import` instead of `require`
