+++ jest.config.js
@@ -8,7 +8,7 @@ module.exports = {
   transform: {
     '^.+\\.(ts|tsx)$': 'ts-jest',
   },
-  testRegex: '(/src/.*\\.(test|spec))\\.(jsx?|tsx?)$',
+  testRegex: '(/test/.*\\.(test|spec))\\.(jsx?|tsx?)$',
   testEnvironment: 'node',
   coverageDirectory: './coverage/',
   collectCoverage: true,
