+++ test/utils-applyMask.test.ts
@@ -1,4 +1,4 @@
-const { applyMask } = require('./utils')
+import { applyMask } from '../src/utils'
 
 describe('applyMask()', () => {
   const items = [
