+++ test/utils-sumElementsByMultipliers.test.ts
@@ -1,4 +1,4 @@
-const { sumElementsByMultipliers } = require('./utils')
+import { sumElementsByMultipliers } from '../src/utils'
 
 describe('sumElementsByMultipliers()', () => {
   test('Tipo do retorno', () => {
