+++ tsconfig.json
@@ -21,5 +21,5 @@
     }
   },
   "exclude": ["./tsconfig.json"],
-  "include": ["./src/**/*"]
+  "include": ["./src/**/*", "./test/**/*"]
 }
